--- packages/main/src/hooks/myPage/subPage/profile_assets_created.js
@@ -1,12 +1,0 @@
-
-const profileAssetCreated = () => {
-
-    let container = document.createElement('div');
-    container.innerHTML += `
-    
-        
-    `
-
-}
-
-export default profileAssetCreated
